+++ controllers/admin/AdminMollieAjaxController.php
@@ -241,7 +241,9 @@ class AdminMollieAjaxController extends ModuleAdminController
             $transactionId = Tools::getValue('transactionId');
             $refundAmount = (float) Tools::getValue('refundAmount') ?: null;
             $orderLineId = Tools::getValue('orderline') ?: null;
+            $orderId = (int) Tools::getValue('orderId') ?: null;
             $quantity = Tools::getValue('quantity') ? (int) Tools::getValue('quantity') : null;
+            $refundType = Tools::getValue('refundType') ?: null;
             if ($quantity !== null && $quantity < 1) {
                 $quantity = null;
             }
@@ -249,7 +251,7 @@ class AdminMollieAjaxController extends ModuleAdminController
             /** @var RefundService $refundService */
             $refundService = $this->module->getService(RefundService::class);
 
-            $status = $refundService->handleRefund($transactionId, $refundAmount, $orderLineId, $quantity);
+            $status = $refundService->handleRefund($transactionId, $refundAmount, $orderLineId, $quantity, $orderId, $refundType);
 
             $this->ajaxRender(json_encode($status));
         } catch (\Throwable $e) {
@@ -345,7 +347,7 @@ class AdminMollieAjaxController extends ModuleAdminController
             $captureService = $this->module->getService(CaptureService::class);
 
             $amount = $captureAmount ? (float) $captureAmount : $order->total_paid;
-            $status = $captureService->handleCapture($transactionId, $amount);
+            $status = $captureService->handleCapture($transactionId, $amount, $orderId);
 
             $this->ajaxRender(json_encode($status));
         } catch (\Throwable $e) {