+++ upgrade/Upgrade-6.4.3.php
@@ -1,0 +1,48 @@
+<?php
+/**
+ * Mollie       https://www.mollie.nl
+ *
+ * @author      Mollie B.V. <info@mollie.nl>
+ * @copyright   Mollie B.V.
+ * @license     https://github.com/mollie/PrestaShop/blob/master/LICENSE.md
+ *
+ * @see        https://github.com/mollie/PrestaShop
+ * @codingStandardsIgnoreStart
+ */
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+/**
+ * @param Mollie $module
+ *
+ * @return bool
+ */
+function upgrade_module_6_4_3($module)
+{
+    try {
+        $columnExists = Db::getInstance()->executeS(
+            'SHOW COLUMNS FROM `' . _DB_PREFIX_ . 'mol_payment_method` LIKE \'is_manual_capture\''
+        );
+
+        if (empty($columnExists)) {
+            Db::getInstance()->execute(
+                'ALTER TABLE `' . _DB_PREFIX_ . 'mol_payment_method` ADD COLUMN `is_manual_capture` TINYINT(1) DEFAULT 0'
+            );
+        }
+
+        return true;
+    } catch (Exception $e) {
+        PrestaShopLogger::addLog(
+            'Mollie module upgrade to 6.4.3 failed: ' . $e->getMessage(),
+            3,
+            $e->getCode(),
+            'Module',
+            $module->id,
+            true
+        );
+
+        return false;
+    }
+}